Setting up routing in Node.js with Connect and Express
Next, I did the basic server setup and routing in Node.js, starting with the HTTP library to run index.js as the main point. Routing is then enhanced with Connect. For more advanced capabilities, Enhanced which was built upon Connect, offers more features and modules. This manages multiple routes in the project. Finally, Express was installed to help with Enhanced.